+++ test_app/app/channels/dexie_channel.rb
@@ -1,0 +1,3 @@
+class DexieChannel < ApplicationCable::Channel
+  include DexieCable
+end